refactor: Improve script and stylesheet loading in index.html
This commit reorganizes the loading of scripts and stylesheets in the `index.html` template. Specifically, it moves the `htmx.org` script and the `app.js` module script to `<script type="module" src="/build/js/app.js" defer></script>` and adds `defer` attribute to the js script to ensure that the DOM is fully loaded before the script is executed. This change enhances the application's performance and modularity.