fix: Remove overzealous dashboard styling
This commit removes excessive styling applied to the dashboard, specifically removing `flex-direction: column` and setting `height: 100%` directly on the `#dashboard` div.  These changes were causing layout issues and unnecessary styling. The changes were made to improve the responsiveness and overall appearance of the dashboard. The `column is-7` class and related styles were also removed to simplify the layout.